requirements.txt
Django==1.3.1
Mezzanine==0.12.5
PIL==1.1.7
South==0.7.3
Werkzeug==0.8.2
distribute==0.6.24
django-easymode==1.0b1
django-extensions==0.7.1
django-modeltranslation==0.3.2
django-rosetta==0.6.3
filebrowser-safe==0.1.3
grappelli-safe==0.1.12
wsgiref==0.1.2
